refactor(run-e2e): sort files
