+++ automation/run-e2e/commands/ci.mjs
@@ -1,0 +1,3 @@
+export async function ci(options) {
+    console.log("Run ci with options:", options);
+}
