+++ automation/run-e2e/commands/dev.mjs
@@ -1,0 +1,3 @@
+export async function dev(options) {
+    console.log("Run dev with options:", options);
+}
