+++ automation/run-e2e/run-dev.mjs
@@ -1,8 +1,8 @@
-import { execSync } from "node:child_process";
-// import { relativeToModule } from "../utils.mjs";
+import { spawnSync } from "node:child_process";
+import { delimiter } from "node:path";
 
 export async function runDev() {
-    console.log("Run dev");
+    console.log("Run e2e as dev");
     // execSync("echo $PATH", {stdio: 'inherit'});
     // const __dirname = fileURLToPath(new URL(".", import.meta.url));
     // const configPath = relativeToModule("cypress.config.js", import.meta.url);
@@ -10,6 +10,5 @@ export async function runDev() {
     const command = [`cypress open`, `--browser chrome`, `--e2e`, `--config-file cypress.config.cjs`].join(" ");
     console.log(command);
 
-
-    execSync(command, { stdio: 'inherit' })
+    execSync(command, { stdio: "inherit" });
 }
